Feat: :rocket: basic authentication on expenses and profits

+++ app/Http/Controllers/Api/ExpensesController.php
@@ -6,20 +6,27 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\ExpensesRequest ;
 use App\Services\ExpensesService;
 use App\Http\Requests\IdRequest;
+use App\Services\UserService;
 
 class ExpensesController extends Controller
 {
     protected $expenseService;
+    protected $UserService;
 
-    public function __construct(ExpensesService $expenseService)
+    public function __construct(ExpensesService $expenseService, UserService $UserService)
     {
         $this->expenseService = $expenseService;
+        $this->UserService = $UserService;
     }
 
     public function create(ExpensesRequest $request)
     {
-        // dd($request);
         $validatedData = $request->validated();
+        $userExists = $this->UserService->checkUserExistence($request['idUser']);
+
+        if ($userExists instanceof \Illuminate\Http\JsonResponse) {
+            return $userExists; // Retorna a resposta de erro caso o usuário não exista
+        }
         $expense = $this->expenseService->createExpense($validatedData);
 
         return response()->json([
@@ -30,6 +37,11 @@ class ExpensesController extends Controller
 
     public function index($idUser)
     {
+        $userExists = $this->UserService->checkUserExistence($idUser);
+
+        if ($userExists instanceof \Illuminate\Http\JsonResponse) {
+            return $userExists; // Retorna a resposta de erro caso o usuário não exista
+        }
         $expenses = $this->expenseService->list($idUser);
 
         return response()->json([
@@ -41,6 +53,11 @@ class ExpensesController extends Controller
     public function update(ExpensesRequest $request, $id)
     {
         $validatedData = $request->validated();
+        $userExists = $this->UserService->checkUserExistence($request['idUser']); 
+
+        if ($userExists instanceof \Illuminate\Http\JsonResponse) {
+            return $userExists; // Retorna a resposta de erro caso o usuário não exista
+        }
         $expense = $this->expenseService->update($id, $validatedData);
 
         return response()->json([
@@ -53,6 +70,11 @@ class ExpensesController extends Controller
     public function destroy(IdRequest $idDelete, $id)
     {
         $idDelete = $idDelete->validated();
+        $userExists = $this->UserService->checkUserExistence($idDelete['idUser']); 
+
+        if ($userExists instanceof \Illuminate\Http\JsonResponse) {
+            return $userExists; // Retorna a resposta de erro caso o usuário não exista
+        }
         $this->expenseService->delete($idDelete,$id);
 
         return response()->json([
