+++ app/Http/Controllers/Api/ProfitsController.php
@@ -6,19 +6,28 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\profitsResquest;
 use App\Services\ProfitsService;
 use App\Http\Requests\IdRequest;
+use App\Services\UserService;
 class ProfitsController extends Controller
 {
     protected $profitService;
+    protected $UserService;
 
-    public function __construct(ProfitsService $profitService)
+    public function __construct(ProfitsService $profitService, UserService $UserService)
     {
         $this->profitService = $profitService;
+        $this->UserService = $UserService;
     }
 
     public function create(profitsResquest $request)
     {
         // Valida os dados da requisição
         $validatedData = $request->validated();
+        // Verifica se o usuário existe
+        $userExists = $this->UserService->checkUserExistence($validatedData['idUser']);
+
+        if ($userExists instanceof \Illuminate\Http\JsonResponse) {
+            return $userExists; // Retorna a resposta de erro caso o usuário não exista
+        }
 
         // Extrai o idUser (assumindo que é parte dos dados validados)
         $idUser = $validatedData['idUser'];
@@ -39,6 +48,13 @@ class ProfitsController extends Controller
     public function index($idUser)
     {
         try {
+
+            $userExists = $this->UserService->checkUserExistence($idUser);
+
+            if ($userExists instanceof \Illuminate\Http\JsonResponse) {
+                return $userExists; // Retorna a resposta de erro caso o usuário não exista
+            }
+
             // Chama o método list passando o idUser para obter a lista de lucros para o usuário
             $profits = $this->profitService->list($idUser);
 
@@ -60,6 +76,12 @@ class ProfitsController extends Controller
     {
         // Valida os dados da requisição
         $validatedData = $request->validated();
+        // Verifica se o usuário existe
+        $userExists = $this->UserService->checkUserExistence($validatedData['idUser']);
+
+        if ($userExists instanceof \Illuminate\Http\JsonResponse) {
+            return $userExists; // Retorna a resposta de erro caso o usuário não exista
+        }
 
         // Chama o método update passando o id do lucro e os dados validados
         $profit = $this->profitService->update($id,$validatedData);
@@ -77,6 +99,11 @@ class ProfitsController extends Controller
         try {
             // Chama o método delete para excluir o lucro com o id fornecido
             $idDelete = $idDelete->validated();
+            $userExists = $this->UserService->checkUserExistence($idDelete['idUser']);
+
+            if ($userExists instanceof \Illuminate\Http\JsonResponse) {
+                return $userExists; // Retorna a resposta de erro caso o usuário não exista
+            }
             $this->profitService->delete($idDelete,$id);
             // Retorna a resposta com sucesso
             return response()->json([
